Move StepButton & StepList to UI/components
