+++ src/features/profile/pages/TrackEmailChange/TrackEmailChange.tsx
@@ -11,12 +11,12 @@ import { useGoBack } from 'features/navigation/useGoBack'
 import { UpdateAppBanner } from 'features/profile/components/Banners/UpdateAppBanner'
 import { Step } from 'features/profile/components/Step/Step'
 import { StepCard, StepCardType } from 'features/profile/components/StepCard/StepCard'
-import { StepList } from 'features/profile/components/StepList/StepList'
 import { getEmailUpdateStep } from 'features/profile/helpers/getEmailUpdateStep'
 import { useEmailUpdateStatus } from 'features/profile/helpers/useEmailUpdateStatus'
 import { useFeatureFlag } from 'libs/firebase/firestore/featureFlags/useFeatureFlag'
 import { RemoteStoreFeatureFlags } from 'libs/firebase/firestore/types'
 import { BackButton } from 'ui/components/headers/BackButton'
+import { StepList } from 'ui/components/StepList/StepList'
 import { BicolorEmailIcon } from 'ui/svg/icons/BicolorEmailIcon'
 import { BicolorNewIcon } from 'ui/svg/icons/BicolorNewIcon'
 import { BicolorPhoneIcon } from 'ui/svg/icons/BicolorPhoneIcon'
